Add server-streaming call
Switch to observer-based handlers. This will allow easier addition for
the other call types, and we can provide helpers for dealing with
observers later.

This also means we can't have RPCs that return `nil`, but I consider
that a feature. The same limitations exist for `core.async` channels.

Finally, we're switching on the method name to determine which
`ServerCall` to use. This is a hack, and we'll need to generalize it.